Update patch set 4
Patch Set 4:

This is a first in a series of refactorings for a larger patch that reduces the work in a number of optimizer passes. The global idea is that each pass will only need to examine the subset of the tree (in most cases is a subset of methods) that were potentially impacted by a previous pass.

The result is that most descendants of JModVisitor will need to track the methods that were modified by the pass. So I refactored the tracking of the current method to JVisitor and the recording of modified methods/fields to JModVisitor (the JModVisitor refactoring is the next patch in the queue).

I see your point and I agree that the patch by itself is not much of an improvement, but I see the added methods (getCurrentMethod() and getCurrentField()) as part of the (inheritance) contract as much as the methods visit() and endVisit().

Patch-set: 4