Report a proper error when the JDT compiler aborts.
That is, print the file and line where the internal error happened, instead of
spurious, nondeterministic errors about being unable to find java/lang/Object
and so on. Once the file is identified it's possible to work around it;
sometimes GWT doesn't need to compile the file at all.

Also added JavaDoc on some methods.

Change-Id: I74397bfa2be4e75ec116f42611197a7fa3953701