Upon selection in a SuggestBox's pop up, focus the text field
After selecting a value from a SuggestBox's suggestion list, the
suggestion list closes and focus snaps back to the body
element. However, users expect the focus to be in the corresponding
ValueBox. So upon the user selecting a value, we set the focus to the
ValueBox per default.

Fixes issue 8051.

Change-Id: I08ba1abf8432ef4ad65c0a6b57983067d66b0d9f