Update patch set 2
Patch Set 2:

(4 comments)

I think they could be part of the public API but it will be more work. We should make sure that the API is what we want to support long-term. It doesn't necessarily have to happen now, but if you feel strongly about it we could do it.

It seems like there are three logically separate changes for SerializableThrowable, StackTraceDeobfuscator, and the format change in Throwable itself, which should have its own external issue since it will be a breaking change for some tools. Could you split the CL's and we can take it from there?

I made a few comments improving SerializableThrowable. However, a larger issue: why does it extend Throwable? If we thought of it as container for a Throwable rather than a drop-in replacement for a Throwable, it seems like some methods would go away and it would be easier to understand?

Patch-set: 2
Reviewer: Gerrit User 5010 <5010@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=0