Update patch set 1
Patch Set 1:

BlackBerry may have fixed the bug in the most recent versions, but it doesn't look like they force users to upgrade in the same way that Chrome does. We see exceptions caused by this bug several hundred times per day on an application loaded from play.google.com.

I would agree that it's not the kind of thing we should try to work around if the fix wasn't so trivial. But it's very easy to fix inside GWT and hard to fix from an application - I would either have to supersource in this entire file or replace all calls to Integer.parseInt with (int) Long.parseLong. And this is being perceived as a GWT bug even though it's BlackBerry's fault, so I think it's worth fixing so long as the cost is negligible.

Patch-set: 1
Reviewer: Gerrit User 5155 <5155@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=0