Update patch set 1
Patch Set 1:

I think this change affects multi-clients.

If the client fails to connect somehow, code was re-launching until it connects and similarly it was continuing launching if it expects multiple clients.
Now it will just hang there. It is hard to guess all implications of this, but it looks like the logic was there for a reason. 
So we should still keep launching module until expected number of clients connect as before.

Patch-set: 1