Update patch set 1
Patch Set 1:

So the JavaDoc for RunStyle.launchModule() says "Requests initial launch of the browser. This should only be called once per instance of RunStyle."  If JUnitShell is calling launchModule() more than once, that sounds like a bug.  I think it's the RunStyle implementations responsibility to make sure launchModule() starts all of the clients if necessary.

Patch-set: 1