Update patch set 1
Patch Set 1:

RunStyle.initialize() returns the number of clients that will be created. RunStyle.launchModule() is supposed to launch all the clients and then return. It looks like most runstyles do this in a loop, and I don't see any code in a RunStyle that would indicate that it was intended to be called multiple times.

It looks like the intention is that notDone() should be called whenever a browser sends us a message or once a second, whichever is shorter, and it isn't supposed to return false until every client has finished executing the current test. However, the code isn't thread-safe and doesn't take into account that an event could arrive in the MessageQueue while it's in the middle of executing. Maybe it should be rewritten so it takes the next message from the message queue as an argument?

For manual mode, notDone() will print out a message once a second or after each browser connects, until all browsers have started.

Patch-set: 1
Reviewer: Gerrit User 5010 <5010@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=0