Update patch set 1
Patch Set 1:

The way it was written made me think that launchModule was intended to be called multiple times but of course it might be just another bug. if you guys feel confident, that is fine with me. 
It would be great if you can put assertions in the code based on the assumptions done here so that at least we will have a better idea if any assumptions are in correct and also would provide documentation.

Patch-set: 1
Reviewer: Gerrit User 5110 <5110@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0