Update patch set 1
Patch Set 1:

(3 comments)

The public API seems mostly okay, except that if we're going to clean things up, I don't think we want to have a lot of subclasses like we did before. We should change the API so that's no longer necessary.

Also, it's surprising that we got this far without unit tests.

Patch-set: 1
Label: Code-Review=0