Update patch set 1
Patch Set 1:

(3 comments)

In a second look, I agree that this class still needs improvements before becoming a public API and it is not a real priority.
What about first moving it to core/server/impl and later we may (or may not) work on it again to make it public?

Patch-set: 1
Label: Code-Review=0