Update patch set 2
Patch Set 2:

If we move it then Googlers might start updating their subclasses to extend the new class, so we'd end up in an in-between state. Perhaps we should wait on committing this change and finish it later? Or is there some part of it that's needed now for your other work?

Patch-set: 2
Reviewer: Gerrit User 5010 <5010@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=0