Update patch set 1
Patch Set 1:

(2 comments)

I thought about the one versus two-class solution and I don't think it's all that big a difference from the user's point of view. You still have to remember to convert a Throwable to a SerializableThrowable in order to safely send it over GWT-RPC, so it seems like any classes containing a Throwable field will still need some special-case code at serialization time (though not at deserialization time).

Another difference is that in the old class, SerializableThrowable.getThrowable() is probably never called in client-side code and ThrowableWithClassName is never instantiated, so the methods that are there only to satisfy the Throwable superclass might get stripped out by the compiler. But it doesn't seem like a big deal.

The other issue is that any code in GWT-RPC that special-cases Throwable might need to be changed, at least for GWT-RPC parameters. I'm not sure about the consequences.

Patch-set: 1
Label: Code-Review=0