Update patch set 1
Patch Set 1:

If we're going to include simple name, how about throwing in Package as well; when obfuscation is on, package can be "" and simpleName = name = obfuscated name.  With obfuscation off, we have to include the package string and simple name, so might as well make them both accessible.

Note: If we do add Package, we'll also have to update the IsolatedAppClassLoader in gwt-dev mode, as it sadly strips package information.  In my reflection service, gwt-dev needs this ugly bit of code to get a package from a class:
     return Thread.currentThread().getContextClassLoader().loadClass(o.getClass().getName()).getPackage();

Patch-set: 1
Reviewer: Gerrit User 5140 <5140@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=0