Changes StackTraceDeobfuscator to abstract and provide various factory methods.
This change will remove the need for extending StackTraceObfusctor most of the 
time as in almost all scenarios it is extended to get the symbol map from jar 
through the class loader (which makes sense).
Also a factory is provided to get the symbol map from the file system. For all
other cases, developers can construct the deobfuscator with a URL path or extend
the class to override openInputStream method.

Change-Id: I562e052caef8da7f3434319cf11b8984bc347fe5