Update patch set 2
Patch Set 2:

Here are the numbers on two sets of 4 runs.

AVG Before: Pre: 267  Compile: 285   Total: 552
AVG After:   Pre: 238  Compile: 310   Total: 548

Not much savings (4 seconds on average). Comparing fastest runs the gains are 10 seconds.

The surprising part is that the serialized precompilation (precompilation.ser) is ~10% smaller with the patch. 

I think the reason is related to an issue I found recently while debugging my latests change, and that is that some nodes, e.g. JMethodCall have targets that are not part of the AST. 

The other finding is that if the optimization is left in the precompilation (as before this patch) the precompilation.ser file is different each time it is run (that is not the case after the patch). A reason for that maybe some iteration on a HashSet or HashMap. I'll have a look at that.

Patch-set: 2
Reviewer: Gerrit User 5105 <5105@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0