Update patch set 6
Patch Set 6:

Choose the "Patch Set" you wan't to compare with in the "Old Version History" dropdown list above the patchsets, and it'll automatically adjust the list of changed files with changes from that new "base". When you're viewing a file diff, you can also switch the left/right pane to a specific patchset using the small links at the top of the panes.

But in the case of the last 2 patchsets, the changes were upstream: patchset 5 was rebased on top of master (along with changes to the parent commit, because the files had changed in the mean time and there were conflicts), and patchset 6 has no difference from patchset 5 besides its parent commit (I just updated the commit message in the parent).

Note: I explicitly put those changes into a "module-deps" topic. It's a branch in my local repo to make it easier (the last commit in the branch –disentangle module dependencies– depends on the previous 3 ones, and there's no easy way to do it without either make them serial in a branch, or having additional merge commits).

I'll make a new round removing the unneeded prefixing for the inner classes before I forget about it.

I don't really know what's best here. I don't think those first changes are controversial, so they can be merged without waiting for the modularization work to settle down, but then they should probably be "reviewed and merged" (individually) in one go, so I can rebase the other commits on master where the merged ones would now be and stop pushing "noise" with patchsets not containing any real change.
The alternative is to wait and only review them when we're ready to start modularizing/mavenizing GWT.
I could also have these commits only in my fork on GitHub, but I'd rather have only commits there about moving files around, and push required changes upstream here and cherry-pick them in the fork (not sure if I'm clear: the idea is to prepare the work here by disentangling things, commits that can be.

In the mean time, all I can do, I'm afraid, is to apologize for the noise (I'd also accept any proposal for an alternative approach to the problem :-) )

Patch-set: 6
Label: Verified=0