Update patch set 4
Patch Set 4:

Looks like we save the bytecode to a map named JdtCompiler.binaryTypes in JdtCompiler.CompileImpl.process(). This map is used to implement INameEnvironmentImpl which we pass to the JdtCompiler for type lookup. It can also read bytecode from generators and falls back to .class files in the compiler's classpath.

It seems like a good source of confusing bugs. For the purposes of this change, actual source code sent to the GWTCompiler takes priority and so @GwtIncompatible will be honored in source. If there are any @GwtIncompatible annotations in the classpath then they wouldn't be stripped, but I think it only matters for the JDT's type checking.

Patch-set: 4