Update patch set 5
Patch Set 5:

This approach does not seem to work for anonymous inner classes. Anonymous inner classes appear in the statement part of the AST as a QualifiedAllocationExpression. The problem is that parsing has two entry points: parse() that build the whole AST and dietParse() that builds the AST minus method bodies but chases references inside method bodies.

Patch-set: 5
Label: Verified=0