Update patch set 8
Patch Set 8:

This feature is so awesome, it's a shame it might have to wait for 2.6.0...  My project manager's reaction was "we've been waiting for this for years!".  

Any hope for a 2.5.2 release with this and anything that didn't quite make it in 2.5.1?  (Sorry to bother you here; it's just it's a big win I'd like to land sooner rather than later; we're doing lots of cross platform work at our shop).

Thanks for this; I'll be using it in trunk on my own projects asap. :)

Patch-set: 8
Reviewer: Gerrit User 5140 <5140@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0