Change RemoteServiceServlet so gwt.codeserver.port works with more subclasses.
In many subclasses of RemoteServiceServlet, the doGetSerializationPolicy()
method is overridden so that it doesn't call super. As a result, setting
gwt.codeserver.port has no effect. To avoid this, I moved the calling code
into getSerializationPolicy(), which is a final method.

A subclass can still disable fetching serialization policies from
the code server by overriding getCodeServerPolicyUrl() to return 0,
but this won't happen by accident.

Change-Id: Ib47db371b78eabb1a139948744279b36aa306bc1