RequestFactoryGenerator generated invalid java classes.
In some cases RequestFactoryGenerator generated classes that implemented the same generic
interface twice with different generic parameters. For example:

  class X extends AbstractRequest<ResultType>
    implements InstanceRequest<BarProxy, ResultType> { ... }

where AbstractRequest is defined in the library as follows:

  class AbstractRequest<X> implements InstanceRequest<BaseProxy, X> {... }

Althogh the code is not valid java the GWT compiler would ignore the error and generate
working code; but newer versions of the JDT reject the code.

The change here is to add a type argument to AbstractRequest:
   AbstractRequest<P extends BaseProxy, X> … implements InstanceRequest<P, X>

Change-Id: Ib9a604ef63f05bbecd8bfed60255f74682607fcc
Review-Link: https://gwt-review.googlesource.com/#/c/2370/