Adds ability to continue testing rest of the module on compile errors.
With this change GWTTestCase will not quit early on module compilation when
the module includes a file that doesn't compile. Before this was causing
NPE with the new code generator for GWTTestCase as the test class was not
found in TypeOracle.

Exiting the code generation early with an error is not a solution because
in that case compiler was not showing the actual compilation error that caused
the class to disappear.

This change is fixing that by using a place holder class if the test class is
not found. Under normal circumstances place holder is not called as the test 
case fails earlier but it is there for sanity check.

Change-Id: Ibe92e8a0a7cb425d2c9fef698664c3a766193cba
Review-Link: https://gwt-review.googlesource.com/#/c/2400/