Update patch set 2
Patch Set 2:

This was previous behavior and I broke it with my code generator change. I wouldn't try hard adding the behavior back however if we fail-fast in codegen as a side effect we can't see the compilation errors :(
On the other hand, we still report the test failures as we hit them and the only case it can continue running other tests is if the test doesn't depend on the not compiling class. If the code is that much unrelated then I would not consider it a fail fast scenario.

Patch-set: 2