Allow @UIHandler annotated methods to be private.
Even though @UIHandler annotated methods were not allowed to 
be private, a bug in JDT 3.4 was dropping the private modifier from
methods that has annotations in parameters and this was causing UIHandler
to work with private methods.

The JDT bug is fixed in JDT 3.8, thus migrating to JDT 3.8 causes 
breakage in many projects that unknowingly depend on the bug being present.

Change-Id: I03c7787ced5a339a74d49ec917d48edebd5dbb8f