Update patch set 2
Patch Set 2:

FWIW, I agree with Thomas that I'd prefer to try to change this into a warning and see if we can fix up code rather than just concede that private methods can be used for @UiHandlers.

Perhaps we make it into a warning for the next major release, and then change it back into an error (as originally intended) for the next one after that?  That should give users plenty of time to adapt.

Patch-set: 2