Update patch set 2
Patch Set 2:

Actually, I encouraged this solution so I will try to answer. 

I think enabling private methods with @UiHandler (and later with UiField) is a reasonable feature not just a workaround. 

While using UiBinder, many times I hoped that was how it was behaving originally - especially with UiFields.
If there were no UiHandler magic then we would have created an anonymous class to delegate to a private handler method instead of exposing the handler method outside. Current behavior (forcing it to be package-visible) I think is just an incident of how generators usually work.

IIRC, this new behavior is also consistent with @Inject in google-gin which again makes sense to me.

Patch-set: 2