Update patch set 2
Patch Set 2:

The surprise comes from violating expectations. I would expect all callers to a private method to be in the same file. And tools assume this too; for example IDEA will mark a method as unused if it private and there are no uses in the same file. So, I might delete the method, then recompile and discover that it is used after all. And I wonder if other tools might be fooled, for example someday there might be a tool to automatically remove all private unused methods in your code, and such tools would need to be modified to look for the UIHandler annotation.

This isn't a big deal today since there is no such tool and there is an annotation, which seems like a sufficient warning to the reader. But I do think it's slightly simpler and more future-proof if "private" means what the Java specification says, not "private except for these special cases that you have to remember."

Patch-set: 2