Update patch set 2
Patch Set 2:

The JsInliner should take care of this case. Methods that are called in only one place (like the *__native one in this case) are inlined. The only drawback is that this happens quite late. 

I'll have a look and make sure it is properly inlined and see if we it makes us miss some optimization opportunities.

Patch-set: 2