Update patch set 2
Patch Set 2:

I agree that we should indirect through JSNI only if the method is actually private. This could be a performance issue in devmode because it adds a round trip between the browser and the JVM. (Is there some way to avoid the round trip?)

I think it's simpler to not change the spec but I will defer to people who use UiBinder more.

Patch-set: 2
Reviewer: Gerrit User 5010 <5010@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0