Reduce O(N^2) memory complexity to O(N)
For long strings with lots of newlines, the current recursive
algorithm results in a bunch of nested call frames, each keeping a
large portion of the original string in memory.  Switching to an
iterative algorithm avoids keeping more than one or two copies of the
input string alive in memory at a time.

Change-Id: I30301a58bad0244a41b08fb534c4cb18b90c9494