Update patch set 3
Patch Set 3:

I've only skimmed, but it looks like changing the format for a dispId would be a wire protocol change for devmode. If we have to deploy all the plugins, that would be a major change.

How much do Elemental users need Dev Mode? I'm thinking Super Dev Mode would be a better fit since the integration with JavaScript is better.

This patch is somewhat of a mess; perhaps we should start over with a simple patch that fails with a clean error when the limits are exceeded, instead of difficult to diagnose errors later.

Patch-set: 3
Reviewer: Gerrit User 5010 <5010@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0