Update patch set 3
Patch Set 3:

Yes, this patch will not be able to fix the actual underlying issue; the best it could do is mask the problem with & 0xffff, which would just lead to indeterminism.

if (dispId & 0xfff > 0) throw new Error("You've got a lot of JSOs there, shouldn't you be using super dev mode?");

The part that is actually breaking (in class Jsni) is actually just looking for the classId; it gets the merged int and then fails when extracting the clsId.  A check for jsos / JavaScriptObject$ beforehand could bypass the overflow.  Thing is, it failed on PotentialElement, and not JavaScriptObject$.  PotentialElement's DispatchClassInfo was filled with methods from JSO$ (80,000+ on skeleton project).

Patch-set: 3