Changes error messaging in devmode.js.
This change makes sure we show the glass pane in dev mode even the error
function is set via gwt:onLoadErrorFn.

If developer has not whitelisted host url, then GWTTestCase will fail when
executed in manual mode using dev mode. However in Chrome it is not clear
why the test case fails. Only message is:
  Failed to load module 'blah'. Please see log for details.

Of course there is no nothing in the logs and dev mode doesn't show the error in
glass pane as it would normally do because gwt:onLoadErrorFn is set in junit.html.

It is not an easy to fix it on junit side as the cause is not communicated back
to error handling function and I don't want to change the contract.

Change-Id: Iee4d0540d843b27623c7b6b9d3702df060a0acbf