Update patch set 1
Patch Set 1:

A lot of people will tell you that putting binaries in a VCS is a Bad Thing™, and 99% of the time they'll be right. That wouldn't be an issue for GWT_TOOLS if we keep the "append only" rule, otherwise it means that cloning the repo also clones all previous versions of the binaries, that many of us will never use. For once, this is something SVN does well: an easy way to synchronize a snapshot of a directory tree.

Maven doesn't play well with dependencies that aren't in a Maven repo (other build tools allow this, not Maven), but it's rather easy to deploy our dependencies in a Maven repo. And for people that don't want Maven to connect to the Internet, you can setup a mirror and have everyone use it in their settings.xml (this is recommended in all organizations anyway). Generally, the mirror downloads from the Internet but you can configure it so that it only hosts artifacts, or possibly allow it to connect from time to time (e.g. when you bootstrap a new build, or you know you'll need new dependencies or new versions of them). That said, if (as?) Google keeps its own non-Maven build, there's no reasons to change the way you deal with 3rd-party libs; the main change is that because it'll be easier for us (using Maven) to add new dependencies or update them, we'll have to try not doing it too often and you'll have to keep an eye on changes on any pom.xml file. You can use then the maven-dependency-plugin to update your GWT_TOOLS with dependencies from Maven (and go through your internal process for updating it): http://maven.apache.org/plugins/maven-dependency-plugin/copy-dependencies-mojo.html

It was Ray who first proposed moving to Maven last year, so maybe he already thought about this (and if not, well, you know who to blame ;-) )

Patch-set: 1