Update patch set 4
Patch Set 4:

-Addressed all javadoc issues.
-Moved back to iterating over child elements.
-added test for mixed child elements

@Brian
I think this is is very helpful to all GWT applications that display some kind of loading indicator. Since this is mostly done with some HTML in the hostpage there should be imho a way to remove such a loading indicator.
RootPanel is a good place to add something like this, instead of having such code lurk around in many different places.

Patch-set: 4
Label: Verified=0