Update patch set 4
Patch Set 4:

(1 comment)

As I noted earlier in my comments, I agree with Brian.

With the added removeAllChildren, to get rid of stuff like loading indicator, all they need to do is this:
 RootPanel.get().getElement().removeAllChildren().

It will definitely help to add a javadoc to RootPanel#clear to point out this use case.

Anyway it is not a big deal either way, so I'm ok with both.

Patch-set: 4
Label: Verified=0