Update patch set 1
Patch Set 1:

One other option (which was my starting point) is just to add generics directly to Composite and ResizeComposite. 
Disadvantages with that are:
1. In a lot of cases generics will not add value but developers will need to provide a type anyway.
2. Old code will give warnings.
3. Change in the ResizeComposite will be a breaking change as pushes the type check to compile time instead of runtime (which is a good thing but a breaking change)

I'm still in-between so would love to hear what you think.

Also note that;
The previous attempt from rjrjr was failed because it was using IsWidget as boundary type which was changing method signatures after type erasure. Because of that it was a breaking change. This one uses Widget as the boundary type so it is backward compatible (Using 'IsWidget' doesn't add much value to Composite anyway).

Patch-set: 1