Fixes a missing flow in DataflowOptimizer that might cause incorrect code.
The DataflowOptimizer was building an incorrect control flow graph in the case
that a catch block ended in a throw (in all paths) and a finally block was 
also present. In this case the graph would be missing the flow from the end
of the catch block to the finally block.

An example snippet that would be optimized incorrectly is

Object o = null;
try {
  if (b) throw new Exception();
    k++;
} catch (Exception e) {
  o = e;
  throw e;
} finally {
  if (o == null) {
    j++;
  }
}

where the optimizer would assume that o is always null at the finally block
and thus transform the code into:

Object o = null;
try {
  if (b) throw new Exception();
    k++;
} catch (Exception e) {
  o = e;
  throw e;
} finally {
  j++;
}

Fixes issue 8115.

Change-Id: I608bd25157eb0c76ac771697ca70756751c10750