Update patch set 1
Patch Set 1:

(3 comments)

Seems okay in principle, though I think we could use review from other widget experts since I'm not an expert.

For the record, I talked to Goktug about this and here's some context:

The reason we're doing this is to make it easier to create panel widgets that you can drop into a UiBinder and also add children to them. This required either not using a Composite (resulting in a messy API) or writing a fair amount of boilerplate, and it wasn't easy to introduce a base class so the boilerplate is in one place. Having the type variable helps to make this clean.

Patch-set: 1
Label: Verified=0