Update patch set 1
Patch Set 1: Code-Review+1

(3 comments)

Thanks for doing this!

To clarify the use case: our project used to build a bunch of URIs as String, which triggered GWT warnings in UI binder when doing: <a href="{field.getUrl()}". So we switched to SafeUri, which now requires adding asString() all over the codebase.

Patch-set: 1
Label: Code-Review=+1