Update patch set 2
Patch Set 2:

Hmm, I don't know how AT deal with it in practice, but <frame longdesc=""> and <q cite=""> are supposed to be navigable. I otherwise agree that maybe not all properties that take URIs need to be updated with a SafeUri overload, though it'd be harmless to have them (except maybe for ScriptElement, which is unsafe by definition).

Also, I had a <object data="javascript:alert('pwnd!')"></object> (built with JS in the inspector) execute and show the alert.

Patch-set: 2
Label: Verified=0