Update patch set 5
Patch Set 5:

JavaScriptException is in the client package and contains JSNI, so it cannot be used on the server (some JVMs will immediately give a link exception if you classload a class with native methods and it can't find the native library for the class).

While currently we leak a ton of client classes to the server and it mostly works for most people if you don't reference the wrong thing, that isn't a good solution and we have been working at removing such references -- we shouldn't add more.

It really needs to be a hard rule that no class in shared ever references any class in client (or server).

Patch-set: 5