Update patch set 5
Patch Set 5:

For this scenarion, I think the class is not loaded in server side so JVM won't give a link exception.

If we are going to put a hard rule on this, we can just remove instanceof check for JavaScriptException (the related code is just a best effort improvement in exception logging with no class metadata so it wouldn't worth the complexity of super-sourcing).

Patch-set: 5