Update patch set 2
Patch Set 2:

It is definitely the deRPC tests at the bottom of RPCSuite.  Commenting them out fixes the test failure.  I'm trying to determine now which tests specifically are at fault, or whether it's all of them.

Maybe we can just move the deRPC tests into a separate test suite?

I'm curious if this is going to be a backwards compatibility issue for users at all.  I notice one internal Google user each for RpcServlet and HybridServiceServlet.  I'll have to check if there test cases are affected by this change at all. :/

Patch-set: 2