Update patch set 2
Patch Set 2:

The tests that were commented out simply extend the other tests that run through a normal RemoteServiceServlet.  The only difference is they use a different .gwt.xml module that loads up extensions of HybridServiceServlet which extends deRPCs RpcServlet.

If legacy support is needed, than a JUnitHostDeRPCImpl that still extends HybridServiceServlet should be used instead.  Given that it's deprecated and being phased out though, this change should likely be made only in the projects that require it.

Additionally, it might save users trouble debugging if the StringIndexOOB is caught, and wrapped in a message that says "You are using the deprecated deRPC RpcServlet instead of the preferred RemoteServiceServlet.".  

The StringIndexOOB is reliably thrown when using a RemoteService interface on a deRPC servlet (unless the payload contains '~'); deRPC scans for a separator char of ~, which is not a special char in standard rpc.

It's actually rather confusing, since both servlets call an RPC.decodeRequest() command, but from two different RPC classes.  c.g.g.rpc.server.SimplePayloadDecoder is where the OOB happens, and is where the message to switch from RpcServlet to RemoteServiceServlet should likely be placed.

Patch-set: 2