Update patch set 4
Patch Set 4:

The issue Goktug is referring to is that code like

    anchor.setHref(null);

used to compile okay, but now it's ambiguous as to whether it should be calling anchor.setHref((String) null) or anchor.setHref((SafeUri) null).  Importing this change into Google caused a handful of projects to start failing to build due to this.

Patch-set: 4