Update patch set 2
Patch Set 2: Code-Review+1

Looks good, but after checking out the tests that are disabled, I've noticed the only difference is the module xml is setting gwt.rpc.hijackLegacyInterface to true so the legacy remote service hijacks the generator to emit deRPC implementations (which appears to not work if JUnitHostImpl is not a deRPC as well).

Unless we are planning to support the legacy deRPC system, we may want to consider deleting the com.google.gwt.rpc.RPCSuite module entirely and making the test cases extend RemoteServiceServlet instead of HybridServiceServlet.  If we are not supporting Hybrid / deRPC, we probably shouldn't leave test servlets that suggests we do support it.

The only servlets other than the ones commented out that extend HybridServiceServlet are from a test for a bug that was fixed seven years ago (in the package "test").
https://code.google.com/p/google-web-toolkit/issues/detail?id=9

Patch-set: 2
Reviewer: Gerrit User 5140 <5140@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=+1