Update patch set 1
Patch Set 1:

This is an ugly hack, but it'll finally allow us to switch our internal testing from Firefox 3.6 to Firefox 17 (ESR).

I don't know why calling .isNull() on a JSONObject instance could return a non-null value, but only on FF17.  I'm wondering if there could be a compiler optimization bug at play?  Or maybe there's a FF17 VM bug that we need to workaround?  (Hopefully not...)

Either way, unless anyone can see a downside to this patch, I'd like to submit it so we can finally drop Firefox 3.6 internally.

Patch-set: 1
Label: Verified=0