Update patch set 2
Patch Set 2: Code-Review+2

Let's merge this then.

Ideally, we'll try to track down the issue later, or assume the issue stale when the next Firefox for Organizations will be out with a fix and/or we replace com.google.gwt.json with a better API (e.g. Elemental), whichever comes first.

BTW, should we report the issue on Mozilla's Bugzilla? Or are we confident that it's entirely our fault (JUnit architecture, compiler output, etc.) that we're "incompatible" with Firefox 17 here?

Patch-set: 2
Label: Code-Review=+2