Update patch set 4
Patch Set 4:

Actually I'm not much concerned about the breaking part, it is an easy fix and there is not so many uses of it.

I'm more concerned about the API. Setting to null is a valid use case; currently there is no other way to revert the element back to its initial state; it is different then setting it to empty or anything else. Most of internal uses was basically using null to reverting the href.

Also If you look at dom.client package, only previous safe setXXX method is the Element#setInner*Safe*Html. This is also inconsistent with that.

I think it is better to change this before people starts using it.

Patch-set: 4