Update patch set 2
Patch Set 2:

That's what I thought it was about; this is great!

Just a note that the getDelegate could possibly break some people that defined a getDelegate method in their Composite subclass. I don't think it'll break many people in practice though, so I'm fine with this (same for the overloaded initWidget, with even less impact I guess).

Patch-set: 2