Update patch set 1
Patch Set 1:

Thanks for the insight John. That was what I was guessing.

The reason I think this is fine without testing is; internally GWT-SDK uses Scheduler much more than the Timer API. So even if there is a leak, keeping this code in Timer will not help much.

Patch-set: 1