Update patch set 1
Patch Set 1: Code-Review-2

(2 comments)

AFAICT, this change is kind of useless without some "negotiation" of the protocol version on the server-side (writing with the same protocol version as the request used), and deferred-binding on the client-side to use protocol version 7 on IE6-7 (where JSON.parse is not available; I think we can safely assume all other browsers have JSON.parse: http://caniuse.com/json).

This requires a huge refactoring of the server code though.

This change, as is (except entirely removing the eval() codepath), will be OK in a year from now though, when we'll definitely remove support for IE6-7 (I though we talked about the end of this year on the SC meeting, which would rather be GWT 2.6, but IIUC Ray & Daniel's I/O talk they said GWT 3.0)

Patch-set: 1
Label: Code-Review=-2