Update patch set 2
Patch Set 2:

Removed the protocol version for now since it will cause a lot of work to change. 

Instead, changed the commit so the only real change is that NaN and Infinity is sent as strings instead as a numbers making the JSON valid. On the client side then JSON.parse can be used to parse the JSON and finally when reading the double values Number(x) is used to make sure the doubles (and floats) are numbers. This should be backward compatible with IE6/7 as well as JSONUtil.safeEval() will fall back to eval() in that case.

Patch-set: 2