Update patch set 2
Patch Set 2:

(1 comment)

As Colin said, this change is not backwards-compatible when transmitting large payloads (which will use array .concat()). Could this be handled dynamically? (look for ").concat(" in the payload and use eval() in this case –could lead to some false positives but that should be rare enough, and relatively harmless–, otherwise use JSON.parse if available or eval() otherwise). As a side note, JsonUtils.safeEval has some overhead in the non-JSON.parse case compared to the bare eval() done in ClientSerializationStreamReader, and even JsonUtils.unsafeEval has a small overhead.

I also wonder if changing the serialization of Infinity and NaN isn't enough to necessitate a bump of protocol version.

If it were just me, I'd drop IE6/7 support as soon as the next version (and IE8 by fall next year, or no later than in 2 years from now), then we could simply remove the .concat() workaround and switch everyone to JSON.parse (and bumping the protocol version to 8). I'd be happy to discuss that again on gwt-steering.

Patch-set: 2